requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==22.0.0
psycopg2-binary==2.9.9
openai==1.55.3
pydantic==2.5.2
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.4
